# route-ai-hub-main/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Handoff Smart Router API","version":"0.1.0"},"paths":{"/route":{"post":{"summary":"Route a clinical request","operationId":"route_request_route_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/events":{"get":{"summary":"List Events","operationId":"list_events_events_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response List Events Events Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/events/stream":{"get":{"summary":"Events Stream","operationId":"events_stream_events_stream_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/fhir/ServiceRequest":{"post":{"summary":"FHIR adapter → /route","operationId":"fhir_service_request_fhir_ServiceRequest_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Fhir Service Request Fhir Servicerequest Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/facilities":{"get":{"summary":"List Facilities","operationId":"list_facilities_facilities_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Facilities Facilities Get"}}}}}}},"/patients/{id}":{"get":{"summary":"Get Patient","operationId":"get_patient_patients__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Patient Patients  Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/metrics":{"get":{"summary":"Metrics","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Metrics Metrics Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RouteRequest":{"properties":{"patient_id":{"type":"string","title":"Patient Id"},"request_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Type"},"department":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"},"urgency":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Urgency"},"location_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Hint"},"free_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Free Text"},"intended_site":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Intended Site"}},"type":"object","required":["patient_id"],"title":"RouteRequest"},"RouteResponse":{"properties":{"route_to_facility_id":{"type":"string","title":"Route To Facility Id"},"route_to_endpoint":{"type":"string","title":"Route To Endpoint"},"confidence":{"type":"number","title":"Confidence"},"decision_status":{"type":"string","title":"Decision Status"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"features_used":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Features Used"},"decision_id":{"type":"string","title":"Decision Id"},"ttl_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Ttl Seconds","default":3600},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"},"resolved_request_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resolved Request Type"},"facility_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Facility Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"hours":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Hours"}},"type":"object","required":["route_to_facility_id","route_to_endpoint","confidence","decision_status","decision_id"],"title":"RouteResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}